COMAAS-459 Fix unread messages counter in Cassandra - Removing CounterIncrementController, legacy functionality regarding only ebayk - Cassandra counter is not re-written by 0 or 1 if the conversation is updated but should update real number of unread messages - Adding more fine-grained methods and more meaningful methods into Cassandra such as: markConversationAsRead, deleteConversation
Reviewed-on: https://gerrit.ecg.so/58818
Tested-by: Gerrit-At-Gitmaster001
Tested-by: Comaas Jenkins <DL-eCG-comaas-dev@ebay.com>

+++ messagecenters/core-messagecenter/src/test/java/com/ecg/messagecenter/persistence/simple/PostBoxTest.java
@@ -56,7 +56,7 @@ public class PostBoxTest {
         ), 180);
 
         postBox.markConversationUnread("convid2", null);
-        assertEquals(1l, postBox.getNewRepliesCounter().getValue().longValue());
+        assertEquals(1, postBox.getNewRepliesCounter().getValue());
         assertTrue(postBox.getUnreadConversations().containsKey("convid2"));
     }
